Confirmed application can transition to a declined state (#25)
In this scenario, someone has cancelled after already registering for
AndConf. This allows the application to transition to a "declined" state
and returns the user to the "applicant" state, making the application
and user states consistent with the behavior of declining prior to
registration. This does NOT delete the attendance record.